+++ auto_route/lib/src/common/common.dart
@@ -1,5 +1,6 @@
 export 'auto_route_annotations.dart';
 export 'auto_route_wrapper.dart';
+export 'param_converter.dart';
 export 'parameters.dart';
 export 'transitions_builders.dart';
 export 'auto_route_observer.dart';